+++ depeche_db/_executor.py
@@ -5,10 +5,11 @@ import threading as _threading
 import time as _time
 from typing import Dict, List, Optional
 
-from ._interfaces import RunOnNotification
+from ._interfaces import FixedTimeBudget, RunOnNotification
 from .tools import PgNotificationListener
 
 
+# TODO add tests
 class Executor:
     """
     Executor is a class that runs handlers on notifications.
@@ -82,7 +83,10 @@ class Executor:
             try:
                 handler = self.handler_queue.get(timeout=0.5)
                 try:
-                    handler()
+                    # TODO make time budget configurable (global and per handler)
+                    # TODO make time budget dependent on pressure (e.g. handler queue length)
+                    # TODO the handler should return if stopped because of time budget, so that we can re-queue it.
+                    handler(budget=FixedTimeBudget(seconds=1))
                 except Exception:
                     self._stop()
                     raise